requirements.txt
asgiref==3.8.1
Cerberus==1.3.7
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
